Safari-609.1.11.1/LayoutTests/fast/dom/attribute-event-listener-errors.html

<body>
	<pre>This test verifies that an attribute event listener error shows the right line number even if the attribute contains newlines.</pre>

	<button id ="1" onclick="&#10;&#10;&#10;error()"></button>
	<button id ="2" onclick="&#10;&#10;&#10;$@_."></button>

<script>
if (window.testRunner)
	testRunner.dumpAsText();

document.getElementById("1").click();
document.getElementById("2").click();
</script>

</body>
