Safari-609.1.11.1/LayoutTests/fast/dom/attribute-event-listener-errors-expected.txt

CONSOLE MESSAGE: line 4: ReferenceError: Can't find variable: error
CONSOLE MESSAGE: line 5: SyntaxError: Invalid character: '@'
This test verifies that an attribute event listener error shows the right line number even if the attribute contains newlines.
  
